use cache instead of current_tip table
