+++ libs/sdk-core/src/persist/mod.rs
@@ -1,5 +1,4 @@
 pub(crate) mod cache;
-pub(crate) mod chain;
 pub(crate) mod channels;
 pub(crate) mod db;
 pub(crate) mod error;
